Chaging To Encoded Execution
Powershell has a feature of executing base64 encoded commands. In order
to have that work the commands need to be base64 encoded with unicode.

If you're on powershell you use this to encode a basic command:

[Convert]::ToBase64String([System.Text.Encoding]::Unicode.GetBytes('Get-ChildItem'))

In order to encode it correctly in python you need to specify the correct
encodings. This is basically:

        encoded_bytes = base64.b64encode(command.encode("utf-16-le"))
        encoded_str = str(encoded_bytes, "utf-8")

It is important to use little-endian here.